# openapi.yaml
openapi: 3.0.3
info:
  title: SalesMind API
  description: Mental fitness platform for sales teams - API for managing user profiles, assessments, skills, drills, and progress tracking
  version: 1.0.0
  contact:
    name: SalesMind Support
    email: support@salesmind.com
servers:
  - url: https://api.salesmind.com/v1
    description: Production server
  - url: https://staging-api.salesmind.com/v1
    description: Staging server
  - url: http://localhost:3000/api
    description: Local development server
...
paths:
  /auth/signup:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Create a new user account with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: SecurePass123!
                full_name:
                  type: string
                  example: John Doe
      responses:
        '201':
          description: User successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Authenticate user
      description: Login with email and password credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: SecurePass123!
      responses:
        '200':
          description: Successfully authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      description: End the current user session
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successfully logged out
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Logged out successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Obtain a new access token using a refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /assessments:
    get:
      tags:
        - Assessments
      summary: Get user's assessments
      description: Retrieve all assessments completed by the user
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          description: Maximum number of assessments to return
          schema:
            type: integer
            default: 10
            maximum: 100
        - name: offset
          in: query
          description: Number of assessments to skip
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Assessments retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Assessment'
                  count:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Assessments
      summary: Submit new assessment
      description: Submit a completed mental skills assessment
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - answers_data
                - calculated_scores
              properties:
                total_questions:
                  type: integer
                  example: 25
                answers_data:
                  type: object
                  description: User's answers to assessment questions
                  example:
                    emotionalResilience1: 2
                    emotionalResilience2: 1
                calculated_scores:
                  type: object
                  description: Calculated skill scores from assessment
                  example:
                    emotional-resilience: 72
                    confidence-building: 45
                assessment_version:
                  type: string
                  example: "1.0"
      responses:
        '201':
          description: Assessment submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assessment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /assessments/{assessmentId}:
    get:
      tags:
        - Assessments
      summary: Get assessment details
      description: Retrieve details of a specific assessment
      security:
        - BearerAuth: []
      parameters:
        - name: assessmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Assessment retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assessment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or signup
  schemas:
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refresh_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          description: Token expiration time in seconds
          example: 3600
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        email:
          type: string
          format: email
          example: user@example.com
        created_at:
          type: string
          format: date-time
...
    Assessment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        completed_at:
          type: string
          format: date-time
        total_questions:
          type: integer
          example: 25
        answers_data:
          type: object
          description: User's answers to assessment questions
        calculated_scores:
          type: object
          description: Calculated skill scores
          example:
            emotional-resilience: 72
            confidence-building: 45
            stress-management: 75
        assessment_version:
          type: string
          example: "1.0"
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Invalid request
        message:
          type: string
          example: The provided data is invalid
        status_code:
          type: integer
          example: 400
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            message: Authentication token is missing or invalid
            status_code: 401
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Not Found
            message: The requested resource does not exist
            status_code: 404